restructure Tools to use autogenerated sidebar
Restructure the docs directory so that Tools are grouped together and they use the Docusaurus autogeneraed sidebar. This should facilitate the automatic inclusion of new tool pages by reducing the number of files that need to be edited.

Also add `index.md` files to tool group directories to include descriptions (place where we will ultimately migrate tool-groups page content to).