reformat odd SVG icons to match standard
These icons had different SVG formatting from the others that resulted in them rendering much larger on the page than others. I fixed this by opening them up in Adobe Illustrator and re-exporting them as SVGs. line counts are much smaller now too. I'm not sure what was the problem but I'm proceeding with these new files that look the same but shrunken to the appropriate size.

Could be that these differences were masked previously and the new package-lock.json update in the last commit resulted in the website losing some dependencies that allowed it to parse the "older" SVGs correctly.